+++ Main/IoC.cs
@@ -60,6 +60,7 @@ using TAOM.Features.CultureConversion;
 using TAOM.Features.MissionDiagnostic;
 using TAOM.Features.CrashReport;
 using TAOM.Features.BattleLoadDiagnostics;
+using TAOM.Features.MapLoadDiagnostics;
 using TAOM.Features.SettlementNameplateFade;
 using TAOM.Features.MenuLinkColors;
 using TAOM.Features.LordSpawnGuard;
@@ -146,6 +147,7 @@ public static class IoC
         MissionDiagnosticIoC.RegisterMissionDiagnosticFeature(container);
         CrashReportIoC.RegisterCrashReportFeature(container);
         BattleLoadDiagnosticsIoC.RegisterBattleLoadDiagnosticsFeature(container);
+        MapLoadDiagnosticsIoC.RegisterMapLoadDiagnosticsFeature(container);
         Features.SaveLoadDiagnostics.SaveLoadDiagnosticsIoC.RegisterSaveLoadDiagnosticsFeature(container);
         Features.BlowDiagnostics.BlowDiagnosticsIoC.RegisterBlowDiagnosticsFeature(container);
         Features.AutoResolveDiagnostics.AutoResolveDiagnosticsIoC.RegisterAutoResolveDiagnosticsFeature(container);